[DB][C-03] 랜드마크 테이블 정규화

+++ Backend/miraclebird/src/main/java/com/ssafy/miraclebird/dto/LandmarkDto.java
@@ -35,7 +35,7 @@ public class LandmarkDto {
 
     private String content;
 
-    private int starForce;
+    private long starForce;
 
     private Boolean selling;
